+++ Client/package.json
@@ -10,7 +10,6 @@
     "@testing-library/user-event": "^12.1.10",
     "@types/jest": "^26.0.15",
     "@types/jscodeshift": "0.7.2",
-    "@types/node": "^12.0.0",
     "@types/prettier": "^2.2.0",
     "@types/react-dom": "^16.9.8",
     "@types/react-redux": "7.1.15",
@@ -62,6 +61,7 @@
     ]
   },
   "devDependencies": {
+    "@types/node": "^12.19.15",
     "@types/react": "^16.14.2"
   }
 }
