+++ Client/src/index.tsx
@@ -1,27 +1,27 @@
 import 'bulmaswatch/superhero/bulmaswatch.min.css';
 import '@fortawesome/fontawesome-free/css/all.min.css';
 import './main.css';
-import ReactDOM from 'react-dom'; 
+import React from 'react';
+import { render } from 'react-dom'; 
 import { Provider } from 'react-redux';
 import { store } from './state';
-import { Header, Navigation, CellList, Signup, Login  } from './components';
-import { useEffect, useState } from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
+import App from './App';
 
-const App = () => {
-    return (
-        <Provider store={store({})}>
+const renderApp = () => {
+    render (
+        <React.StrictMode>
+               <Provider store={store({})}>
             <Router>
-                <Navigation />
-                <Header />
-                <Switch>
-                    <Route exact path="/" component={CellList} />
-                    <Route exact path="/signup" component={Signup} />
-                    <Route exact path="/login" component={Login} />
-                </Switch>
+                <App />
             </Router>
         </Provider>
+        </React.StrictMode>,
+        document.getElementById('root')
     );
 };
 
-ReactDOM.render(<App />, document.querySelector('#root'));
+// @ts-ignore
+if (process.env.NODE_ENV !== "production" && module.hot) module.hot.accept("./App", renderApp);
+
+renderApp();
